+++ r/R/port4me.R
@@ -36,6 +36,19 @@ parse_ports <- function(ports) {
   ports
 }
 
+port4me_include <- function() {
+  include <- NULL
+
+  for (name in c("PORT4ME_INCLUDE", "PORT4ME_INCLUDE_SITE")) {
+    arg <- Sys.getenv(name, "")
+    ports <- parse_ports(arg)
+    include <- c(include, ports)
+  }
+  include <- unique(include)
+  
+  include
+}
+
 port4me_exclude <- function() {
   exclude <- NULL
 
@@ -55,7 +68,7 @@ port4me_skip <- function() {
   skip
 }
 
-port4me <- function(user = port4me_user(), tool = port4me_tool(), exclude = port4me_exclude(), skip = port4me_skip(), list = NULL, test = NULL, max_tries = 1000L, must_work = TRUE) {
+port4me <- function(user = port4me_user(), tool = port4me_tool(), include = port4me_include(), exclude = port4me_exclude(), skip = port4me_skip(), list = NULL, test = NULL, max_tries = 1000L, must_work = TRUE) {
   stopifnot(length(user) == 1L, is.character(user), !is.na(user))
   stopifnot(is.null(tool) || is.character(tool), !anyNA(tool))
   if (!is.null(list)) {
@@ -63,6 +76,8 @@ port4me <- function(user = port4me_user(), tool = port4me_tool(), exclude = port
   }
   stopifnot(length(max_tries) == 1L, is.numeric(max_tries), !is.na(max_tries), max_tries > 0, is.finite(max_tries))
   max_tries <- as.integer(max_tries)
+  if (is.character(include)) include <- parse_ports(include)
+  stopifnot(is.numeric(include), !anyNA(include), all(include > 0), all(include <= 65535))
   if (is.character(exclude)) exclude <- parse_ports(exclude)
   stopifnot(is.numeric(exclude), !anyNA(exclude), all(exclude > 0), all(exclude <= 65535))
   stopifnot(length(skip) == 1L, is.numeric(skip), !is.na(skip), skip >= 0, is.finite(skip), skip < max_tries)
@@ -76,20 +91,28 @@ port4me <- function(user = port4me_user(), tool = port4me_tool(), exclude = port
 
   lcg_set_seed(port4me_seed(user = user, tool = tool))
 
-  if (!is.null(list)) {
-    return(vapply(seq_len(list), FUN.VALUE = NA_integer_,
-                    FUN = function(kk) lcg_port()))
-  } else if (!is.null(test)) {
+  if (!is.null(test)) {
     return(is_port_free(test))
   }
 
+  ports <- integer(0)
   count <- 0L
   while (count <= max_tries) {
-    port <- lcg_port()
+    if (length(include) > 0) {
+      port <- include[1]
+      include <- include[-1]
+    } else {
+      port <- lcg_port()
+    }
     if (port %in% exclude) next
     count <- count + 1L
     if (count <= skip) next
-    if (is_port_free(port)) return(port)
+    if (is.null(list)) {
+      if (is_port_free(port)) return(port)
+    } else {
+      ports <- c(ports, port)
+      if (length(ports) == list) return(ports)
+    }
   }
 
   if (must_work) stop("Failed to find a free TCP port")