+++ NEWS.md
@@ -1,6 +1,8 @@
 # Version (development version)
 
-* ...
+## New Features
+
+* Add `port4me --include=<ports>`.
 
 
 # Version 0.1.0 [2022-07-24]